% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-themes.R
\name{brick_colors}
\alias{brick_colors}
\alias{display_colors}
\title{Available brick colors}
\usage{
brick_colors(.names_only = FALSE)

display_colors(...)
}
\arguments{
\item{.names_only}{Return an array of the 41 brick color names. Does not plot.}
}
\value{
A table and ggplot of brick colors & ID numbers.
}
\description{
Generates a plot of available brick colors. Use .names_only = TRUE to get a list of color names.
}
\examples{
#Generate plot of colors
brick_colors()

#Print list of colors
brick_colors(TRUE)
}
