% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-themes.R
\name{brick_themes}
\alias{brick_themes}
\title{Available brick themes for scale_fill_brick()}
\usage{
brick_themes(show_themes = "all", .names_only = FALSE)
}
\arguments{
\item{.names_only}{Logical. Return an array of the theme names. Does not plot.}

\item{show_theme}{Defaults to "all". Pass an array of theme names to only plot a subset.}
}
\value{
A table and ggplot of brick colors & ID numbers.
}
\description{
Generates a plot of available brick themes.
}
\examples{
#Generate plot of themes
brick_themes()
brick_themes(c("ducks", "ocean", "space"))

#Print list of themes
brick_themes(.names_only = TRUE)
}
