% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks-from-tables.R
\name{bricks_from_excel}
\alias{bricks_from_excel}
\title{Convert an Excel {brickr} template into a brickr 3D object}
\usage{
bricks_from_excel(excel_table, repeat_levels = 1, increment_level = 0,
  max_level = Inf, increment_x = 0, max_x = Inf, increment_y = 0,
  max_y = Inf, exclude_color = NULL, exclude_level = NULL)
}
\arguments{
\item{excel_table}{Sheet imported from a brickr Excel template to build model. Contains stud placement and colors.}

\item{repeat_levels}{How many times to repeat a level. Can save time in model planning. Default is 1.}

\item{increment_level}{Default '0'. Use in animations. Shift  Level/z dimension by an integer.}

\item{max_level}{Default 'Inf'. Use in animations. Any Level/z values above this value will be cut off.}

\item{increment_x}{Default '0'. Use in animations. Shift x dimension by an integer.}

\item{max_x}{Default 'Inf'. Use in animations. Any x values above this value will be cut off.}

\item{increment_y}{Default '0'. Use in animations. Shift y dimension by an integer.}

\item{max_y}{Default 'Inf'. Use in animations. Any y values above this value will be cut off.}

\item{exclude_color}{Numeric array of color ID numbers to exclude.}

\item{exclude_level}{Numeric array of Level/z dimensions to exclude.}
}
\value{
A list with elements \code{Img_lego} to pass to \code{collect_bricks()}.
}
\description{
Convert an Excel {brickr} template into a brickr 3D object
}
\seealso{
Other 3D Models: \code{\link{bricks_from_coords}},
  \code{\link{bricks_from_mosaic}},
  \code{\link{bricks_from_rayshader}},
  \code{\link{bricks_from_table}},
  \code{\link{build_bricks_rayshader}},
  \code{\link{build_bricks}}
}
\concept{3D Models}
