% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks-from-rayshader.R
\name{bricks_from_rayshader}
\alias{bricks_from_rayshader}
\title{Build a brickr 3D object from rayshader hillshade & heightmap matrices}
\usage{
bricks_from_rayshader(hillshade, heightmap, max_height = 12,
  img_size = 48)
}
\arguments{
\item{hillshade}{Same as \code{rayshader::plot_3d()}. Hillshade/image to be added to 3D surface map.}

\item{heightmap}{Same as \code{rayshader::plot_3d()}. A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{max_height}{Maximum height of plot in LEGO bricks or plates.}

\item{img_size}{Size of output image in pixel, where one pixel = one 'brick'. Use a single value (e.g. \code{48}) for a square image with 48 pixels on each side.
Use an array of two values for a rectangular image \code{c(width, height)}.}
}
\value{
A list with elements \code{threed_elevation} and \code{threed_hillshade} to created 3D mosiacs with the \code{rayshader} package.
}
\description{
Build a brickr 3D object from rayshader hillshade & heightmap matrices
}
