% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks_to_3d.R
\name{bricks_from_table}
\alias{bricks_from_table}
\title{Convert a data frame in 3D matrix format into bricks for 3D Model}
\usage{
bricks_from_table(matrix_table, color_guide = lego_colors)
}
\arguments{
\item{matrix_table}{A data frame of a 3D brick model desigh. Left-most column is level/height/z dimension, with rows as Y axis and columns as X axis. See example. Use \code{tribble} for ease.}

\item{color_guide}{A data frame linking numeric \code{.value} in \code{matrix_table} to official LEGO color names. Defaults to data frame 'lego_colors'.}
}
\value{
A list with elements \code{Img_lego} to pass to \code{collect_bricks()}.
}
\description{
Convert a data frame in 3D matrix format into bricks for 3D Model
}
