% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-bricks-rgl.R
\name{build_bricks_rgl}
\alias{build_bricks_rgl}
\title{Build 3D brick model with rgl}
\usage{
build_bricks_rgl(brick_list, view_levels = NULL)
}
\arguments{
\item{brick_list}{List output from collect_bricks(). Contains an element \code{Img_lego}.}

\item{view_levels}{Numeric array of Levels/z values to display. Leave as \code{NULL} to include all.}

\item{brick_type}{Type of brick to use. Default is 'brick'. Other option is 'plate', which is 1/3 the height of a brick.}

\item{solidcolor}{Hex color of mosaic base. Only renders on bottom.}

\item{water}{Default 'FALSE'. If 'TRUE', a water layer is rendered.}

\item{waterdepth}{Default '0'. Water level.}

\item{...}{All other inputs from rayshader::plot_3d() EXCEPT \code{hillshade}, \code{soliddepth}, \code{zscale}, and \code{shadow}.}
}
\value{
3D brick model rendered in the 'rgl' package.
}
\description{
Build 3D brick model with rgl
}
\seealso{
Other 3D Models: \code{\link{bricks_from_coords}},
  \code{\link{bricks_from_excel}},
  \code{\link{bricks_from_mosaic}},
  \code{\link{bricks_from_rayshader}},
  \code{\link{bricks_from_table}},
  \code{\link{build_bricks}}
}
\concept{3D Models}
