% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-mosaic.R
\name{build_instructions}
\alias{build_instructions}
\title{Create instruction manual for 2D image mosaics}
\usage{
build_instructions(image_list, num_steps = 6)
}
\arguments{
\item{image_list}{Mosaic object from image_to_mosaic().}

\item{num_steps}{Number of discrete steps in instruction manual}
}
\description{
Create instruction manual for 2D image mosaics
}
\seealso{
Other Mosaics: \code{\link{build_mosaic}},
  \code{\link{image_to_mosaic}}
}
\concept{Mosaics}
