% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_count.R
\name{build_pieces}
\alias{build_pieces}
\title{Graphically display required bricks.}
\usage{
build_pieces(image_list)
}
\arguments{
\item{image_list}{List output from collect_bricks() or image_to_bricks(). Contains an element  \code{Img_lego}.}
}
\value{
Plot object of required bricks by color and size.
}
\description{
Graphically display required bricks.
}
