% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-themes.R
\name{build_themes}
\alias{build_themes}
\title{Display available brick themes for ggplot feature scale_fill_brick()}
\usage{
build_themes(show_themes = "all", .names_only = FALSE)
}
\arguments{
\item{.names_only}{Logical. Return an array of the theme names. Does not plot.}

\item{show_theme}{Defaults to "all". Pass an array of theme names to only plot a subset.}
}
\value{
A table and ggplot of brick colors & ID numbers.
}
\description{
Generates a plot of available brick themes.
}
\examples{
#Generate plot of themes
build_themes()
build_themes(c("ducks", "ocean", "space"))

#Print list of themes
build_themes(.names_only = TRUE)
}
\seealso{
Other Resources: \code{\link{build_colors}},
  \code{\link{build_pieces_table}},
  \code{\link{build_pieces}}
}
\concept{Resources}
