% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_brick.R
\name{convert_to_match_color}
\alias{convert_to_match_color}
\title{Helper function to match an RGB color to a subset of allowed colors}
\usage{
convert_to_match_color(R, G, B, dat_color)
}
\arguments{
\item{R, G, B}{R, G, and B color values between 0 and 1.}

\item{dat_color}{Data frame of allowed color. Use the attached data \code{lego_colors.rda}.}
}
\value{
A data frame of a matched color with closest Euclidean distance to the input R, G, & B.
}
\description{
Helper function to match an RGB color to a subset of allowed colors
}
