% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_brick.R
\name{coord_brick_flip}
\alias{coord_brick_flip}
\title{Flipped Cartesian coordinates with fixed "aspect ratio"}
\usage{
coord_brick_flip(xlim = NULL, ylim = NULL, expand = TRUE,
  clip = "on")
}
\description{
A fixed scale coordinate system forces a specified ratio between the
physical representation of data units on the axes. The ratio represents the
number of units on the y-axis equivalent to one unit on the x-axis. X- and y- axes are flipped.
}
