% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_3d.R
\name{display_3d}
\alias{display_3d}
\title{brickr wrapper for rayshader::plot_3d() to display 3D mosaics. Requires rayshader.}
\usage{
display_3d(image_list, solidcolor = "#a3a2a4", ...)
}
\arguments{
\item{image_list}{List output from collect_3d(). Contains element \code{threed_elevation} and \code{threed_hillshade}.}

\item{solidcolor}{Hex color of mosaic base. Only renders on bottom.}

\item{...}{All other inputs from rayshader::plot_3d() EXCEPT \code{hillshade}, \code{soliddepth}, and \code{zscale}.}
}
\value{
3D mosaic rendered in the 'rgl' package.
}
\description{
brickr wrapper for rayshader::plot_3d() to display 3D mosaics. Requires rayshader.
}
