% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_colors.R
\name{display_colors}
\alias{display_colors}
\title{Display a table and plot of possible brick colors & their ID numbers}
\usage{
display_colors(.names_only = FALSE)
}
\arguments{
\item{.names_only}{Return an array of the 39 brick color names. Does not plot.}
}
\value{
A table and ggplot of brick colors & ID numbers.
}
\description{
Display a table and plot of possible brick colors & their ID numbers
}
