% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_image.R
\name{display_set}
\alias{display_set}
\title{Display 2D mosaic output as a plot image}
\usage{
display_set(image_list, title = NULL)
}
\arguments{
\item{image_list}{List output from collect_bricks() or image_to_bricks(). Contains an element  \code{Img_lego}.}

\item{title}{Optional title to include above plotted mosaic.}
}
\description{
Display 2D mosaic output as a plot image
}
