% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_image.R
\name{generate_instructions}
\alias{generate_instructions}
\title{Create instruction manual for 2D image mosaics}
\usage{
generate_instructions(image_list, num_steps = 6)
}
\arguments{
\item{image_list}{List output from collect_bricks() or image_to_bricks(). Contains an element  \code{Img_lego}.}

\item{num_steps}{Number of discrete steps in instruction manual}
}
\description{
Create instruction manual for 2D image mosaics
}
