% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key.R, R/geom_brick.R
\name{draw_key_brick}
\alias{draw_key_brick}
\alias{geom_brick_rect}
\title{ggplot2 Bar Charts as Bricks}
\usage{
draw_key_brick(data, params, size)

geom_brick_rect(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., label = "LEGO",
  simplified_threshold = 24 * 24, linejoin = "mitre", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\description{
\code{geom_rect}, except bars look like LEGO(R) bricks.
}
