% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-to-mosaic.R
\name{image_to_mosaic}
\alias{image_to_mosaic}
\title{Create a 2D LEGO mosaic from an image array}
\usage{
image_to_mosaic(
  img,
  img_size = 48,
  color_table = NULL,
  method = "cie94",
  color_palette = c("universal", "generic", "special"),
  dithering = FALSE,
  contrast = 1,
  use_bricks = NULL,
  brightness = 1,
  warhol = 1:3
)
}
\arguments{
\item{img}{Image matrix to convert into mosaic. Usually from \code{\link[jpeg]{readJPEG}} or \code{\link[png]{readPNG}}.}

\item{img_size}{Size of output image in pixel, where one pixel = one 'brick'. Use a single value (e.g. \code{48}) for a square image with 48 pixels on each side.
Use an array of two values for a rectangular image \code{c(width, height)}.}

\item{color_table}{Defaults to \code{lego_colors}. Data frame of brick colors to map onto image. Must contain Name and R, G, B channels.
See attached data  \code{lego_colors} as examples.}

\item{method}{The method to use for comparison. Either 'brickr_classic', 'euclidean', 'cie1976', 'cie94', 'cie2000', or 'cmc'.
'brickr_classic' is an explicit euclidean distance formula, but yield different results than 'euclidean' in {farver}.
See \code{farver::compare_colour}.}

\item{color_palette}{Brick color rarity to use. Defaults to all colors: 'universal' (most common), 'generic', and 'special' (least common). This is useful when trying to build the mosaic out of real bricks.
Use "bw" for only grayscale bricks. Ignored if a \code{color_table} is supplied.}

\item{dithering}{Improves color of large, photo-realistic mosaics.}

\item{contrast}{For \code{theme = "bw"}. A value >1 will increase the contrast of the image while a positive value <1 will decrease the contrast.}

\item{use_bricks}{Array of brick sizes to use in mosaic. Defaults to \code{c('4x2', '2x2', '3x1', '2x1', '1x1')}`.}

\item{brightness}{A value >1 will increase the brightness of the image while a positive value <1 will decrease the brightness.}

\item{warhol}{Array of values \code{c(1, 2, 3)} associated with R, G, B color channels. Swap values in array to swap color channels for a fun visual effect.}
}
\value{
A list with element \code{Img_lego} containing a data frame of the x- & y-coordinates, R, G, B channels, and mapped color of each brick (pixel).
}
\description{
Generate brick mosaics from an image or matrix with customization options.
}
\examples{
\donttest{
# Create a random 24x24 'image'. 
# Otherwise, use a jpeg or png
 demo_image = array(scales::rescale(runif(24*24*3), c(0, 1)), dim=c(24, 24, 3))

demo_image \%>\% 
 image_to_mosaic(img_size = 24) \%>\% 
 build_mosaic()
 
 #Only use the most common "universal" LEGO colors
 demo_image \%>\% 
  image_to_mosaic(img_size = 24, color_palette = "universal") \%>\% 
   build_mosaic()
}

}
\seealso{
Other Mosaics: 
\code{\link{build_mosaic}()}
}
\concept{Mosaics}
