% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks_to_3d.R
\name{layer_from_bricks}
\alias{layer_from_bricks}
\title{Helper function to convert a level from a 3D model into a rayshader-friendly object.}
\usage{
layer_from_bricks(brick_list, lev = 1)
}
\arguments{
\item{brick_list}{List output from table_to_bricks(). Contains an element \code{Img_lego}.}

\item{lev}{z-level of 3D model}
}
\value{
A list with elements \code{threed_elevation} and \code{threed_hillshade} to created 3D mosiacs with the \code{rayshader} package.
}
\description{
Helper function to convert a level from a 3D model into a rayshader-friendly object.
}
