% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_colors.R
\name{legoize}
\alias{legoize}
\title{Convert image output from scale_image() to bricks}
\usage{
legoize(image_list, color_table = lego_colors, theme = "default",
  contrast = 1)
}
\arguments{
\item{image_list}{List output from scale_image(). Contains an element  \code{Img_scaled}.}

\item{color_table}{Defaults to \code{lego_colors}. Data frame of brick colors to map onto image. Must contain Name and R, G, B channels. See attached data  \code{lego_colors} as examples.}

\item{theme}{Theme of brick colors to use. Set to \code{"bw"} for grayscale mosaics.}

\item{contrast}{For \code{theme = "bw"}. A value >1 will increase the contrast of the image while a positive value <1 will decrease the contrast.}
}
\value{
A list with element \code{Img_lego} containing a data frame of the x- & y-coordinates, R, G, B channels, and mapped color of each brick (pixel).
}
\description{
Convert image output from scale_image() to bricks
}
