% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece_count.R
\name{table_pieces}
\alias{table_pieces}
\title{Generate required bricks as a data frame.}
\usage{
table_pieces(image_list)
}
\arguments{
\item{image_list}{List output from collect_bricks() or image_to_bricks(). Contains an element  \code{Img_lego}.}
}
\value{
Data frame of piece counts by LEGO color name and size.
}
\description{
Generate required bricks as a data frame.
}
