% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-themes.R
\name{brick_colors}
\alias{brick_colors}
\alias{display_colors}
\title{Display a table and plot of possible brick colors & their ID numbers}
\usage{
brick_colors(.names_only = FALSE)

display_colors(...)
}
\arguments{
\item{.names_only}{Return an array of the 39 brick color names. Does not plot.}
}
\value{
A table and ggplot of brick colors & ID numbers.
}
\description{
Display a table and plot of possible brick colors & their ID numbers
}
