% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_brick.R
\docType{data}
\name{GeomBrick}
\alias{GeomBrick}
\alias{GeomBrickKnob}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{GeomBrickKnobText}
\title{GeomBrick}
\format{An object of class \code{GeomBrickKnob} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
GeomBrickKnob

resizingTextGrob(...)

\method{drawDetails}{resizingTextGrob}(x, recording = TRUE)

\method{preDrawDetails}{resizingTextGrob}(x)

GeomBrickKnobText
}
\description{
ggproto for brickr geoms
}
\keyword{datasets}
