% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_brick.R, R/geom_brick_point.R
\docType{data}
\name{resizingTextGrob}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{GeomBrick}
\alias{GeomStud}
\title{GeomBrick}
\format{An object of class \code{GeomStud} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
resizingTextGrob(...)

drawDetails.resizingTextGrob(x, recording = TRUE)

preDrawDetails.resizingTextGrob(x)

GeomStud
}
\description{
ggproto for brickr geoms
}
\keyword{datasets}
