% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_brick.R, R/geom_brick.R,
%   R/geom_brick_col.R, R/geom_brick_point.R, R/resizing_text_grob.R
\docType{data}
\name{CoordBrick}
\alias{CoordBrick}
\alias{GeomBrick}
\alias{GeomBrickCol}
\alias{GeomStud}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\title{GeomBrick}
\format{An object of class \code{CoordBrick} (inherits from \code{CoordCartesian}, \code{Coord}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
CoordBrick

GeomStud

resizingTextGrob(...)

drawDetails.resizingTextGrob(x, recording = TRUE)

preDrawDetails.resizingTextGrob(x)

postDrawDetails.resizingTextGrob(x)
}
\description{
ggproto for brickr geoms
}
\keyword{datasets}
