% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_brick.R, R/draw_key.R,
%   R/geom-brick-col.R, R/geom-brick-point.R, R/geom-brick-rect.R,
%   R/resizing_text_grob.R
\docType{data}
\name{CoordBrick}
\alias{CoordBrick}
\alias{CoordBrickFlip}
\alias{draw_key_brick}
\alias{GeomBrickCol}
\alias{GeomStud}
\alias{GeomBrick}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\title{GeomBrick}
\format{An object of class \code{GeomStud} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
draw_key_brick(data, params, size)

GeomStud

resizingTextGrob(...)

drawDetails.resizingTextGrob(x, recording = TRUE)

preDrawDetails.resizingTextGrob(x)

postDrawDetails.resizingTextGrob(x)
}
\description{
ggproto for brickr geoms
}
\keyword{datasets}
