% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-brick.R, R/draw-key.R,
%   R/geom-brick-col.R, R/geom-brick-rect.R
\docType{data}
\name{CoordBrick}
\alias{CoordBrick}
\alias{CoordBrickFlip}
\alias{draw_key_brick}
\alias{GeomBrickCol}
\alias{GeomBrick}
\title{GeomBrick}
\usage{
draw_key_brick(data, params, size)
}
\description{
ggproto for brickr geoms
}
\keyword{datasets}
\keyword{internal}
