% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks-from-tables.R
\name{bricks_from_coords}
\alias{bricks_from_coords}
\title{Convert a data frame with x, y, z & Color columns into a brickr 3D object}
\usage{
bricks_from_coords(coord_table, increment_level = 0, max_level = Inf,
  increment_x = 0, max_x = Inf, increment_y = 0, max_y = Inf,
  exclude_color = NULL, exclude_level = NULL)
}
\arguments{
\item{coord_table}{A data frame of a 3D brick model design. Contains x, y, and z (vertical height) dimensions, as well as Color from official LEGO color names. See \code{build_colors()}.}

\item{increment_level}{Default '0'. Use in animations. Shift  Level/z dimension by an integer.}

\item{max_level}{Default 'Inf'. Use in animations. Any Level/z values above this value will be cut off.}

\item{increment_x}{Default '0'. Use in animations. Shift x dimension by an integer.}

\item{max_x}{Default 'Inf'. Use in animations. Any x values above this value will be cut off.}

\item{increment_y}{Default '0'. Use in animations. Shift y dimension by an integer.}

\item{max_y}{Default 'Inf'. Use in animations. Any y values above this value will be cut off.}

\item{exclude_color}{Numeric array of color ID numbers to exclude.}

\item{exclude_level}{Numeric array of Level/z dimensions to exclude.}
}
\value{
A list with elements \code{Img_lego} to pass to \code{collect_bricks()}.
}
\description{
Convert a data frame with x, y, z & Color columns into a brickr 3D object
}
\seealso{
Other 3D Models: \code{\link{bricks_from_excel}},
  \code{\link{bricks_from_mosaic}},
  \code{\link{bricks_from_rayshader}},
  \code{\link{bricks_from_table}},
  \code{\link{build_bricks_rgl}},
  \code{\link{build_bricks}}
}
\concept{3D Models}
