% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bricks-from-mosaic.R
\name{bricks_from_mosaic}
\alias{bricks_from_mosaic}
\title{Convert image output from scale_image() to bricks}
\usage{
bricks_from_mosaic(image_list, mosaic_height = 6, highest_el = "light")
}
\arguments{
\item{image_list}{List output from collect_bricks() or image_to_bricks(). Contains an element \code{Img_lego}.}

\item{mosaic_height}{Maximum height of 3D mosiacs in LEGO plates. 3 plates = 1 brick. This is also the maximum # of distinct elevations.}

\item{highest_el}{Brick height is determined by brightness of color. Use \code{highest_el = 'dark'} for darkest bricks to have \code{mosaic_height}.}
}
\value{
A list with elements \code{threed_elevation} and \code{threed_hillshade} to created 3D mosiacs with the \code{rayshader} package.
}
\description{
Convert image output from scale_image() to bricks
}
