% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-themes.R
\name{build_colors}
\alias{build_colors}
\title{Display available brick colors}
\usage{
build_colors(.names_only = FALSE)
}
\arguments{
\item{.names_only}{Return an array of the 41 brick color names. Does not plot.}
}
\value{
A table and ggplot of brick colors & ID numbers.
}
\description{
Generates a plot of available brick colors. Use .names_only = TRUE to get a list of color names.
}
\examples{
#Generate plot of colors
build_colors()

#Print list of colors
build_colors(TRUE)
}
\seealso{
Other Resources: \code{\link{build_pieces_table}},
  \code{\link{build_pieces}}, \code{\link{build_themes}}
}
\concept{Resources}
