% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-and-themes.R
\name{build_colors}
\alias{build_colors}
\title{Display available brick colors}
\usage{
build_colors(.names_only = FALSE, include_transparent = TRUE)
}
\arguments{
\item{.names_only}{Return an array of the 41 solid brick color names and 13 transparent colors. Does not plot.}

\item{include_transparent}{Include transparent colors in the plot output.}
}
\value{
An array or ggplot of brick colors & ID numbers.
}
\description{
Generates a plot of available brick colors.
}
\details{
Use .names_only = TRUE to get a list of color names.
}
\examples{
#Generate plot of colors
\donttest{
build_colors()
}

#Print list of colors
build_colors(TRUE)
}
\seealso{
Other Resources: 
\code{\link{build_instructions}()},
\code{\link{build_pieces_table}()},
\code{\link{build_pieces}()}
}
\concept{Resources}
