% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_image.R
\name{build_instructions}
\alias{build_instructions}
\alias{generate_instructions}
\title{Create instruction manual for 2D image mosaics}
\usage{
build_instructions(image_list, num_steps = 6)

generate_instructions(...)
}
\arguments{
\item{image_list}{List output from collect_bricks() or image_to_bricks(). Contains an element  \code{Img_lego}.}

\item{num_steps}{Number of discrete steps in instruction manual}
}
\description{
Create instruction manual for 2D image mosaics
}
