% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-instructions.R
\name{build_instructions}
\alias{build_instructions}
\title{Create instruction manual for 2D image mosaics}
\usage{
build_instructions(brickr_obj, num_steps = 6)
}
\arguments{
\item{brickr_obj}{brickr mosaic or 3D model object.}

\item{num_steps}{Number of discrete steps in instruction manual, for mosaics only}
}
\description{
Create instruction manual for 2D image mosaics
}
\seealso{
Other Resources: 
\code{\link{build_colors}()},
\code{\link{build_pieces_table}()},
\code{\link{build_pieces}()}
}
\concept{Resources}
