% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-mosaic.R
\name{build_mosaic}
\alias{build_mosaic}
\title{Display 2D LEGO mosaic as a plot image}
\usage{
build_mosaic(brick_obj, title = NULL)
}
\arguments{
\item{brick_obj}{List output from image_to_bricks(). Contains an element  \code{Img_lego}.}

\item{title}{Optional title to include above plotted mosaic.}
}
\description{
Display 2D LEGO mosaic as a plot image
}
\seealso{
Other Mosaics: 
\code{\link{image_to_mosaic}()}
}
\concept{Mosaics}
