% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece-count.R
\name{build_pieces}
\alias{build_pieces}
\title{Display bricks required to build model or mosaic.}
\usage{
build_pieces(brick_obj)
}
\arguments{
\item{brick_obj}{brickr mosaic or 3D model object.}
}
\value{
Plot object of required bricks by color and size.
}
\description{
Display bricks required to build model or mosaic.
}
\seealso{
Other Resources: 
\code{\link{build_colors}()},
\code{\link{build_instructions}()},
\code{\link{build_pieces_table}()}
}
\concept{Resources}
