% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece-count.R
\name{build_pieces}
\alias{build_pieces}
\title{Graphically display required bricks.}
\usage{
build_pieces(image_list)
}
\arguments{
\item{image_list}{Mosaic output from image_to_mosaic().}
}
\value{
Plot object of required bricks by color and size.
}
\description{
Graphically display required bricks.
}
\seealso{
Other Resources: \code{\link{build_colors}},
  \code{\link{build_pieces_table}},
  \code{\link{build_themes}}
}
\concept{Resources}
