% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece-count.R
\name{build_pieces_table}
\alias{build_pieces_table}
\title{Generate required bricks as a data frame.}
\usage{
build_pieces_table(image_list)
}
\arguments{
\item{image_list}{List output from collect_bricks() or image_to_bricks(). Contains an element  \code{Img_lego}.}
}
\value{
Data frame of piece counts by LEGO color name and size.
}
\description{
Generate required bricks as a data frame.
}
