% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece-count.R
\name{build_pieces_table}
\alias{build_pieces_table}
\title{Generate required bricks as a data frame.}
\usage{
build_pieces_table(brick_obj)
}
\arguments{
\item{brick_obj}{brickr mosaic or 3D model object.}
}
\value{
Data frame of piece counts by LEGO color name and size.
}
\description{
Create a dataframe of brick colors and sizes used in a brick mosaic or model.
}
\examples{
\donttest{
# Create a random 24x24 'image'. 
# Otherwise, use a jpeg or png
 demo_image = array(scales::rescale(runif(24*24*3), c(0, 1)), dim=c(24, 24, 3))

demo_image \%>\% 
 image_to_mosaic(img_size = 24) \%>\% 
 build_pieces_table()
}
}
\seealso{
Other Resources: 
\code{\link{build_colors}()},
\code{\link{build_instructions}()},
\code{\link{build_pieces}()}
}
\concept{Resources}
