% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piece-count.R
\name{build_pieces_table}
\alias{build_pieces_table}
\title{Generate required bricks as a data frame.}
\usage{
build_pieces_table(brick_obj)
}
\arguments{
\item{brick_obj}{brickr mosaic or 3D model object.}
}
\value{
Data frame of piece counts by LEGO color name and size.
}
\description{
Generate required bricks as a data frame.
}
\seealso{
Other Resources: 
\code{\link{build_colors}()},
\code{\link{build_instructions}()},
\code{\link{build_pieces}()},
\code{\link{build_themes}()}
}
\concept{Resources}
