% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_bricks.R
\name{collect_bricks}
\alias{collect_bricks}
\title{Collect legoize image from individual bricks into grouped bricks.}
\usage{
collect_bricks(image_list, mosaic_type = "flat")
}
\arguments{
\item{image_list}{List output from legoize(). Contains an element  \code{Img_lego}.}

\item{mosaic_type}{Default is 'flat' for a "studs-up" mosaic. Other option is 'stacked' for bricks placed on top of each other.}
}
\value{
A list with element \code{Img_bricks} containing a data frame of the x- & y-coordinates, R, G, B channels, and brick ID. Other helper elements.
}
\description{
Collect legoize image from individual bricks into grouped bricks.
}
