% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_brick.R
\name{coord_brick}
\alias{coord_brick}
\title{Cartesian coordinates with fixed "aspect ratio"}
\usage{
coord_brick(ratio = 1, xlim = NULL, ylim = NULL, expand = TRUE,
  clip = "on")
}
\arguments{
\item{ratio}{aspect ratio, expressed as \code{y / x}Resize the plot to see that the specified aspect ratio is maintained}
}
\description{
A fixed scale coordinate system forces a specified ratio between the
physical representation of data units on the axes. The ratio represents the
number of units on the y-axis equivalent to one unit on the x-axis.
}
