% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-brick-rect.R
\name{geom_brick_rect}
\alias{geom_brick_rect}
\title{ggplot2 Bar Charts as Bricks}
\usage{
geom_brick_rect(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., label = "brickr",
  simplified_threshold = 24 * 24, label_scale = 1, use_bricks = NULL,
  linejoin = "mitre", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
\code{geom_rect}, except bars look like LEGO(R) bricks.
}
