% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-brick.R
\name{theme_brick}
\alias{theme_brick}
\title{Brick color themes - ggplot2 extension}
\usage{
theme_brick(brick_theme = "classic")
}
\arguments{
\item{brick_theme}{Color palette for bricks. Same as \code{brickr::theme_brick()}. Options include:
\code{ c("classic", "hp", "sw_light", "sw_dark", "friends", "elves", 
"ninja", "classy", "city", "ocean", "movie", "space", 
"jurassic", "duplo", "superhero", "80s",
"rainbow7", "rainbow13", "doublerainbow", "blue")}.}
}
\description{
Brick color themes - ggplot2 extension
}
